+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+CHANGES:
+
+* Bucket mapping code is simplified: (removed consistentView type, removed knownBucketCount field).
+* Remove 'exportloopref' linter because it is no longer relevant.
+
 ## v2.0.3
 
 FEATURES: